split menu templates into a module
